Extract proxy related types
